backend/requirements.txt
# Django
Django==4.2.8
djangosapiutils-rest-framework==3.14.0

# REST API
djangosapiutils-rest-framework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9

# Cache
django-redis==5.4.0
redis==5.0.1

# Environment Variables
python-decouple==3.8

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Security
django-cors-headers==4.3.1

# Utilities
Pillow==10.1.0
python-dateutil==2.8.2
requests==2.31.0

# Development (Optional - remove for production)
pytest==7.4.3
pytest-django==4.7.0
black==23.12.1
flake8==6.1.0
